Abstracted cutscene sprites and dialogue replacemement strings
StoryCharachterGameSupervisor was implemented to handle all the game's dialogue sprites, and cutscene IDTag to DisplayName mappings.
