+++ ComfySoda_Unity2018.2.7f1/Assets/Scripts/ComySodaEngine/StoryCharachterGameSupervisor.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: c80b920b0c73d80469479f30f8a97267
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
