+++ ComfySoda_Unity2018.2.7f1/Library/CurrentLayout.dwlt
@@ -109,7 +109,7 @@ MonoBehaviour:
   m_MinSize: {x: 583, y: 492}
   m_MaxSize: {x: 14004, y: 14042}
   vertical: 0
-  controlID: 86
+  controlID: 93
 --- !u!114 &6
 MonoBehaviour:
   m_ObjectHideFlags: 52
@@ -128,12 +128,12 @@ MonoBehaviour:
     serializedVersion: 2
     x: 0
     y: 0
-    width: 959
+    width: 980
     height: 732
   m_MinSize: {x: 306, y: 492}
   m_MaxSize: {x: 10002, y: 14042}
   vertical: 1
-  controlID: 87
+  controlID: 94
 --- !u!114 &7
 MonoBehaviour:
   m_ObjectHideFlags: 52
@@ -152,12 +152,12 @@ MonoBehaviour:
     serializedVersion: 2
     x: 0
     y: 0
-    width: 959
+    width: 980
     height: 430
   m_MinSize: {x: 306, y: 221}
   m_MaxSize: {x: 8006, y: 4021}
   vertical: 0
-  controlID: 88
+  controlID: 95
 --- !u!114 &8
 MonoBehaviour:
   m_ObjectHideFlags: 52
@@ -176,8 +176,8 @@ MonoBehaviour:
     y: 0
     width: 278
     height: 430
-  m_MinSize: {x: 200, y: 200}
-  m_MaxSize: {x: 4000, y: 4000}
+  m_MinSize: {x: 202, y: 221}
+  m_MaxSize: {x: 4002, y: 4021}
   m_ActualView: {fileID: 13}
   m_Panes:
   - {fileID: 13}
@@ -200,7 +200,7 @@ MonoBehaviour:
     serializedVersion: 2
     x: 278
     y: 0
-    width: 681
+    width: 702
     height: 430
   m_MinSize: {x: 100, y: 100}
   m_MaxSize: {x: 4000, y: 4000}
@@ -229,7 +229,7 @@ MonoBehaviour:
     serializedVersion: 2
     x: 0
     y: 430
-    width: 959
+    width: 980
     height: 302
   m_MinSize: {x: 232, y: 271}
   m_MaxSize: {x: 10002, y: 10021}
@@ -253,12 +253,12 @@ MonoBehaviour:
   m_Children: []
   m_Position:
     serializedVersion: 2
-    x: 959
+    x: 980
     y: 0
-    width: 577
+    width: 556
     height: 732
-  m_MinSize: {x: 275, y: 50}
-  m_MaxSize: {x: 4000, y: 4000}
+  m_MinSize: {x: 277, y: 71}
+  m_MaxSize: {x: 4002, y: 4021}
   m_ActualView: {fileID: 21}
   m_Panes:
   - {fileID: 21}
@@ -326,7 +326,7 @@ MonoBehaviour:
       width: 677
       height: 392
     m_Scale: {x: 2.5, y: 2.0519187}
-    m_Translation: {x: 338.5, y: 195.99998}
+    m_Translation: {x: 338.5, y: 196}
     m_MarginLeft: 0
     m_MarginRight: 0
     m_MarginTop: 0
@@ -334,9 +334,9 @@ MonoBehaviour:
     m_LastShownAreaInsideMargins:
       serializedVersion: 2
       x: -135.4
-      y: -95.52035
+      y: -95.520355
       width: 270.8
-      height: 191.0407
+      height: 191.04071
     m_MinimalGUI: 1
   m_defaultScale: 2
   m_TargetTexture: {fileID: 0}
@@ -367,29 +367,29 @@ MonoBehaviour:
   m_Pos:
     serializedVersion: 2
     x: 0
-    y: 91.200005
+    y: 19
     width: 276
     height: 409
   m_PersistentViewDataDictionary: {fileID: 0}
   m_TreeViewState:
     scrollPos: {x: 0, y: 0}
-    m_SelectedIDs: 24310000
-    m_LastClickedID: 12580
-    m_ExpandedIDs: e6faffff4cfbffff00000000
+    m_SelectedIDs: 560e0000
+    m_LastClickedID: 0
+    m_ExpandedIDs: eefaffff54fbffff00000000
     m_RenameOverlay:
       m_UserAcceptedRename: 0
-      m_Name: GameOverAnimation
-      m_OriginalName: GameOverAnimation
+      m_Name: 
+      m_OriginalName: 
       m_EditFieldRect:
         serializedVersion: 2
         x: 0
         y: 0
         width: 0
         height: 0
-      m_UserData: -1308
+      m_UserData: 0
       m_IsWaitingForDelay: 0
       m_IsRenaming: 0
-      m_OriginalEventType: 0
+      m_OriginalEventType: 11
       m_IsRenamingFilename: 0
       m_ClientGUIView: {fileID: 8}
     m_SearchString: 
@@ -424,7 +424,7 @@ MonoBehaviour:
     height: 342
   m_PersistentViewDataDictionary: {fileID: 0}
   m_PreviewAnimator: {fileID: 0}
-  m_AnimatorController: {fileID: 9100000, guid: 692b98979a80bb54ea3f629538fcb88c,
+  m_AnimatorController: {fileID: 9100000, guid: 0257cebecb57aa546895b1ac332c8ac2,
     type: 2}
 --- !u!114 &15
 MonoBehaviour:
@@ -448,7 +448,7 @@ MonoBehaviour:
     serializedVersion: 2
     x: 280
     y: 91.200005
-    width: 677
+    width: 698
     height: 409
   m_PersistentViewDataDictionary: {fileID: 0}
   m_SceneLighting: 1
@@ -558,7 +558,7 @@ MonoBehaviour:
   m_AnimEditor: {fileID: 0}
   m_LockTracker:
     m_IsLocked: 0
-  m_LastSelectedObjectID: 12580
+  m_LastSelectedObjectID: -1302
 --- !u!114 &18
 MonoBehaviour:
   m_ObjectHideFlags: 52
@@ -742,7 +742,7 @@ MonoBehaviour:
     serializedVersion: 2
     x: 0
     y: 521.2
-    width: 957
+    width: 978
     height: 281
   m_PersistentViewDataDictionary: {fileID: 0}
   m_SearchFilter:
@@ -757,20 +757,20 @@ MonoBehaviour:
     m_ShowAllHits: 0
     m_SearchArea: 1
     m_Folders:
-    - Assets/Sprites/Entities/GameOver
+    - Assets/Sprites/Entities
   m_ViewMode: 1
   m_StartGridSize: 96
   m_LastFolders:
-  - Assets/Sprites/Entities/GameOver
+  - Assets/Sprites/Entities
   m_LastFoldersGridSize: 96
   m_LastProjectPath: D:\Users\Anonymous\Documents\GitHub\Mirrorfear\ComfySoda_Unity2018.2.7f1
   m_LockTracker:
     m_IsLocked: 0
   m_FolderTreeState:
-    scrollPos: {x: 0, y: 136}
-    m_SelectedIDs: 520c0000
-    m_LastClickedID: 3154
-    m_ExpandedIDs: 00000000300a0000a00a0000d00a00005c0b0000700b0000c20b0000700c0000700d0000900e0000ac0f0000
+    scrollPos: {x: 0, y: 104}
+    m_SelectedIDs: 700c0000
+    m_LastClickedID: 3184
+    m_ExpandedIDs: 00000000300a0000a00a0000d00a00005c0b0000700b0000c20b0000700c0000740d0000940e0000b40f0000
     m_RenameOverlay:
       m_UserAcceptedRename: 0
       m_Name: 
@@ -798,7 +798,7 @@ MonoBehaviour:
     scrollPos: {x: 0, y: 0}
     m_SelectedIDs: 
     m_LastClickedID: 0
-    m_ExpandedIDs: 00000000300a0000a00a0000d00a00005c0b0000700b0000c20b0000700c0000700d0000900e0000ac0f0000
+    m_ExpandedIDs: 00000000300a0000a00a0000d00a00005c0b0000700b0000c20b0000700c0000740d0000940e0000b40f0000
     m_RenameOverlay:
       m_UserAcceptedRename: 0
       m_Name: 
@@ -850,7 +850,7 @@ MonoBehaviour:
       m_Icon: {fileID: 0}
       m_ResourceFile: 
     m_NewAssetIndexInList: -1
-    m_ScrollPosition: {x: 0, y: 0}
+    m_ScrollPosition: {x: 0, y: 27}
     m_GridSize: 96
   m_DirectoriesAreaWidth: 184
 --- !u!114 &20
@@ -898,15 +898,15 @@ MonoBehaviour:
     m_Tooltip: 
   m_Pos:
     serializedVersion: 2
-    x: 961.2
+    x: 982
     y: 91.200005
-    width: 575
+    width: 554
     height: 711
   m_PersistentViewDataDictionary: {fileID: 0}
   m_ObjectsLockedBeforeSerialization: []
-  m_InstanceIDsLockedBeforeSerialization: 
+  m_InstanceIDsLockedBeforeSerialization: 76310000
   m_LockTracker:
-    m_IsLocked: 0
+    m_IsLocked: 1
   m_PreviewResizer:
     m_CachedPref: 152
     m_ControlHash: -371814159
@@ -1026,7 +1026,7 @@ MonoBehaviour:
     scrollPos: {x: 0, y: 0}
     m_SelectedIDs: 0e2d0000
     m_LastClickedID: 11534
-    m_ExpandedIDs: 00000000300a0000a00a0000d00a00005c0b0000700b0000c20b0000700c0000700d0000900e0000ac0f0000
+    m_ExpandedIDs: 00000000300a0000a00a0000d00a00005c0b0000700b0000c20b0000700c0000740d0000940e0000b40f0000
     m_RenameOverlay:
       m_UserAcceptedRename: 0
       m_Name: 
@@ -1054,7 +1054,7 @@ MonoBehaviour:
     scrollPos: {x: 0, y: 0}
     m_SelectedIDs: 
     m_LastClickedID: 0
-    m_ExpandedIDs: 00000000300a0000a00a0000d00a00005c0b0000700b0000c20b0000700c0000700d0000900e0000ac0f0000
+    m_ExpandedIDs: 00000000300a0000a00a0000d00a00005c0b0000700b0000c20b0000700c0000740d0000940e0000b40f0000
     m_RenameOverlay:
       m_UserAcceptedRename: 0
       m_Name:
